Protect against HashMap rehash spikes in UnresolvedSamples
While adding support for in-kernel unwinding with sample rates as high
as 4KHz, sample dropping became an issue. Looking closer using a
`perf record` of compiler running on 16 CPU, 4KHz samples, rehashing was
taking over 260ms in the worst case (see "Stack Chart" and highlighting
rehash): https://share.firefox.dev/3Y32wsT

This changes so that UnresolvedStacks uses a BTreeMap for the interning
storage, which has a much more bounded O(log n) worst case insertion.
An LRU hash map serves on top as an O(1) fast path.

The new implementation is roughly on par and perhaps slightly faster,
and processing time spikes are almost gone:
https://share.firefox.dev/4pMM9wy

Footnote, the `stacks` array itself also suffers from O(n) realloc
behavior, but that doesn't seem to be problematic *yet*, so I'll leave
this for when it's found to be a problem in the future.